wrote the first few lines for you <3
I'll leave the rest for you to figure out, but this one should be a good starting place to get back into the rhythm of things, my first solution to this has 4 lines of code, why don't you see if you can match or even beat that? I think you should be able to with what I taught you.
If you need help with anything talk to me ok?